fix: consolidate feature flags UI and stop polling on error
- Replace feature-flags-section.tsx with batch save/discard pattern
  (local state, unsaved-change indicators, mark-for-reset)
- Delete duplicate feature-flags-settings.tsx (dead code)
- Stop refetchInterval polling when query is in error state to avoid
  wasting resources when Unleash returns 503
